extract phx setup and config from test_helper; add nav_helpers
